【bug=165】前台登录过期验证、登录前单个投票点击提醒登录

+++ app/src/utils/userLogout.js
@@ -1,9 +1,11 @@
 'use strict'
 import Goto from '../utils/goto'
 const targetPath = '/users/login'
+import {message} from 'antd'
 export default () => {
-  window.localStorage.clear()
-  if (window.location.href.match(/^list/)) {
-    Goto(targetPath)
+  if(window.localStorage.getItem('user.token')){
+    window.localStorage.clear()
+    message.warning('登录失效，请重新登录')
+    setTimeout(Goto(targetPath),5000)
   }
 }
